Code refactoring and restructuring
