+++ Microsoft.Toolkit/Diagnostics/Internals/ThrowHelper.Guard.IO.cs
