+++ Microsoft.Toolkit/Diagnostics/Internals/ThrowHelper.Guard.String.cs
