+++ Microsoft.Toolkit/Diagnostics/Internals/ThrowHelper.Guard.Tasks.cs
