+++ Microsoft.Toolkit/Diagnostics/ThrowHelper.cs
@@ -3,10 +3,11 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
+using System.ComponentModel;
 using System.Diagnostics.CodeAnalysis;
-using System.Diagnostics.Contracts;
-using System.Runtime.CompilerServices;
-using Microsoft.Toolkit.Extensions;
+using System.IO;
+using System.Runtime.InteropServices;
+using System.Threading;
 
 #nullable enable
 
@@ -18,214 +19,667 @@ namespace Microsoft.Toolkit.Diagnostics
     public static partial class ThrowHelper
     {
         /// <summary>
-        /// Returns a formatted representation of the input value.
+        /// Throws a new <see cref="ArrayTypeMismatchException"/>.
         /// </summary>
-        /// <param name="obj">The input <see cref="object"/> to format.</param>
-        /// <returns>A formatted representation of <paramref name="obj"/> to display in error messages.</returns>
-        [Pure]
-        private static string ToAssertString(this object? obj)
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="ArrayTypeMismatchException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowArrayTypeMismatchException(string message)
         {
-            return obj switch
-            {
-                string _ => $"\"{obj}\"",
-                null => "null",
-                _ => $"<{obj}>"
-            };
+            throw new ArrayTypeMismatchException(message);
         }
 
         /// <summary>
-        /// Throws an <see cref="ArgumentException"/> when <see cref="Guard.IsNull{T}(T,string)"/> (where <typeparamref name="T"/> is <see langword="class"/>) fails.
+        /// Throws a new <see cref="ArrayTypeMismatchException"/>.
         /// </summary>
-        /// <typeparam name="T">The type of the input value.</typeparam>
-        [MethodImpl(MethodImplOptions.NoInlining)]
+        /// <param name="message">The message to include in the exception.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="ArrayTypeMismatchException">Thrown with the specified parameters.</exception>
         [DoesNotReturn]
-        internal static void ThrowArgumentExceptionForIsNull<T>(T value, string name)
-            where T : class
+        public static void ThrowArrayTypeMismatchException(string message, Exception innerException)
         {
-            ThrowArgumentException(name, $"Parameter {name.ToAssertString()} ({typeof(T).ToTypeString()}) must be null, was {value.ToAssertString()} ({value.GetType().ToTypeString()})");
+            throw new ArrayTypeMismatchException(message, innerException);
         }
 
         /// <summary>
-        /// Throws an <see cref="ArgumentException"/> when <see cref="Guard.IsNull{T}(T,string)"/> (where <typeparamref name="T"/> is <see langword="struct"/>) fails.
+        /// Throws a new <see cref="ArgumentException"/>.
         /// </summary>
-        /// <typeparam name="T">The type of the input value.</typeparam>
-        [MethodImpl(MethodImplOptions.NoInlining)]
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="ArgumentException">Thrown with the specified parameter.</exception>
         [DoesNotReturn]
-        internal static void ThrowArgumentExceptionForIsNull<T>(T? value, string name)
-            where T : struct
+        public static void ThrowArgumentException(string message)
         {
-            ThrowArgumentException(name, $"Parameter {name.ToAssertString()} ({typeof(T?).ToTypeString()}) must be null, was {value.ToAssertString()} ({typeof(T).ToTypeString()})");
+            throw new ArgumentException(message);
         }
 
         /// <summary>
-        /// Throws an <see cref="ArgumentNullException"/> when <see cref="Guard.IsNotNull{T}(T,string)"/> fails.
+        /// Throws a new <see cref="ArgumentException"/>.
         /// </summary>
-        /// <typeparam name="T">The type of the input value.</typeparam>
-        [MethodImpl(MethodImplOptions.NoInlining)]
+        /// <param name="message">The message to include in the exception.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="ArgumentException">Thrown with the specified parameters.</exception>
         [DoesNotReturn]
-        internal static void ThrowArgumentNullExceptionForIsNotNull<T>(string name)
+        public static void ThrowArgumentException(string message, Exception innerException)
         {
-            ThrowArgumentNullException(name, $"Parameter {name.ToAssertString()} ({typeof(T).ToTypeString()}) must be not null)");
+            throw new ArgumentException(message, innerException);
         }
 
         /// <summary>
-        /// Throws an <see cref="ArgumentException"/> when <see cref="Guard.IsOfType{T}"/> fails.
+        /// Throws a new <see cref="ArgumentException"/>.
         /// </summary>
-        /// <typeparam name="T">The type of the input value.</typeparam>
-        [MethodImpl(MethodImplOptions.NoInlining)]
+        /// <param name="name">The argument name.</param>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="ArgumentException">Thrown with the specified parameters.</exception>
         [DoesNotReturn]
-        internal static void ThrowArgumentExceptionForIsOfType<T>(object value, string name)
+        public static void ThrowArgumentException(string name, string message)
         {
-            ThrowArgumentException(name, $"Parameter {name.ToAssertString()} must be of type {typeof(T).ToTypeString()}, was {value.GetType().ToTypeString()}");
+            throw new ArgumentException(message, name);
         }
 
         /// <summary>
-        /// Throws an <see cref="ArgumentException"/> when <see cref="Guard.IsNotOfType{T}"/> fails.
+        /// Throws a new <see cref="ArgumentException"/>.
         /// </summary>
-        /// <typeparam name="T">The type of the input value.</typeparam>
-        [MethodImpl(MethodImplOptions.NoInlining)]
+        /// <param name="name">The argument name.</param>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="ArgumentException">Thrown with the specified parameters.</exception>
         [DoesNotReturn]
-        internal static void ThrowArgumentExceptionForIsNotOfType<T>(object value, string name)
+        public static void ThrowArgumentException(string name, string message, Exception innerException)
         {
-            ThrowArgumentException(name, $"Parameter {name.ToAssertString()} must not be of type {typeof(T).ToTypeString()}, was {value.GetType().ToTypeString()}");
+            throw new ArgumentException(message, name, innerException);
         }
 
         /// <summary>
-        /// Throws an <see cref="ArgumentException"/> when <see cref="Guard.IsOfType"/> fails.
+        /// Throws a new <see cref="ArgumentNullException"/>.
         /// </summary>
-        [MethodImpl(MethodImplOptions.NoInlining)]
+        /// <param name="name">The argument name.</param>
+        /// <exception cref="ArgumentNullException">Thrown with the specified parameter.</exception>
         [DoesNotReturn]
-        internal static void ThrowArgumentExceptionForIsOfType(object value, Type type, string name)
+        public static void ThrowArgumentNullException(string name)
         {
-            ThrowArgumentException(name, $"Parameter {name.ToAssertString()} must be of type {type.ToTypeString()}, was {value.GetType().ToTypeString()}");
+            throw new ArgumentNullException(name);
         }
 
         /// <summary>
-        /// Throws an <see cref="ArgumentException"/> when <see cref="Guard.IsNotOfType"/> fails.
+        /// Throws a new <see cref="ArgumentNullException"/>.
         /// </summary>
-        [MethodImpl(MethodImplOptions.NoInlining)]
+        /// <param name="name">The argument name.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="ArgumentNullException">Thrown with the specified parameters.</exception>
         [DoesNotReturn]
-        internal static void ThrowArgumentExceptionForIsNotOfType(object value, Type type, string name)
+        public static void ThrowArgumentNullException(string name, Exception innerException)
         {
-            ThrowArgumentException(name, $"Parameter {name.ToAssertString()} must not be of type {type.ToTypeString()}, was {value.GetType().ToTypeString()}");
+            throw new ArgumentNullException(name, innerException);
         }
 
         /// <summary>
-        /// Throws an <see cref="ArgumentException"/> when <see cref="Guard.IsAssignableToType{T}"/> fails.
+        /// Throws a new <see cref="ArgumentNullException"/>.
         /// </summary>
-        /// <typeparam name="T">The type being checked against.</typeparam>
-        [MethodImpl(MethodImplOptions.NoInlining)]
+        /// <param name="name">The argument name.</param>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="ArgumentNullException">Thrown with the specified parameters.</exception>
         [DoesNotReturn]
-        internal static void ThrowArgumentExceptionForIsAssignableToType<T>(object value, string name)
+        public static void ThrowArgumentNullException(string name, string message)
         {
-            ThrowArgumentException(name, $"Parameter {name.ToAssertString()} must be assignable to type {typeof(T).ToTypeString()}, was {value.GetType().ToTypeString()}");
+            throw new ArgumentNullException(name, message);
         }
 
         /// <summary>
-        /// Throws an <see cref="ArgumentException"/> when <see cref="Guard.IsNotAssignableToType{T}"/> fails.
+        /// Throws a new <see cref="ArgumentOutOfRangeException"/>.
         /// </summary>
-        /// <typeparam name="T">The type being checked against.</typeparam>
-        [MethodImpl(MethodImplOptions.NoInlining)]
+        /// <param name="name">The argument name.</param>
+        /// <exception cref="ArgumentOutOfRangeException">Thrown with the specified parameter.</exception>
         [DoesNotReturn]
-        internal static void ThrowArgumentExceptionForIsNotAssignableToType<T>(object value, string name)
+        public static void ThrowArgumentOutOfRangeException(string name)
         {
-            ThrowArgumentException(name, $"Parameter {name.ToAssertString()} must not be assignable to type {typeof(T).ToTypeString()}, was {value.GetType().ToTypeString()}");
+            throw new ArgumentOutOfRangeException(name);
         }
 
         /// <summary>
-        /// Throws an <see cref="ArgumentException"/> when <see cref="Guard.IsAssignableToType"/> fails.
+        /// Throws a new <see cref="ArgumentOutOfRangeException"/>.
         /// </summary>
-        [MethodImpl(MethodImplOptions.NoInlining)]
+        /// <param name="name">The argument name.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="ArgumentOutOfRangeException">Thrown with the specified parameter.</exception>
         [DoesNotReturn]
-        internal static void ThrowArgumentExceptionForIsAssignableToType(object value, Type type, string name)
+        public static void ThrowArgumentOutOfRangeException(string name, Exception innerException)
         {
-            ThrowArgumentException(name, $"Parameter {name.ToAssertString()} must be assignable to type {type.ToTypeString()}, was {value.GetType().ToTypeString()}");
+            throw new ArgumentOutOfRangeException(name, innerException);
         }
 
         /// <summary>
-        /// Throws an <see cref="ArgumentException"/> when <see cref="Guard.IsAssignableToType"/> fails.
+        /// Throws a new <see cref="ArgumentOutOfRangeException"/>.
         /// </summary>
-        [MethodImpl(MethodImplOptions.NoInlining)]
+        /// <param name="name">The argument name.</param>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="ArgumentOutOfRangeException">Thrown with the specified parameters.</exception>
         [DoesNotReturn]
-        internal static void ThrowArgumentExceptionForIsNotAssignableToType(object value, Type type, string name)
+        public static void ThrowArgumentOutOfRangeException(string name, string message)
         {
-            ThrowArgumentException(name, $"Parameter {name.ToAssertString()} must not be assignable to type {type.ToTypeString()}, was {value.GetType().ToTypeString()}");
+            throw new ArgumentOutOfRangeException(name, message);
         }
 
         /// <summary>
-        /// Throws an <see cref="ArgumentException"/> when <see cref="Guard.IsBitwiseEqualTo{T}"/> fails.
+        /// Throws a new <see cref="ArgumentOutOfRangeException"/>.
         /// </summary>
-        /// <typeparam name="T">The type of input values being compared.</typeparam>
-        [MethodImpl(MethodImplOptions.NoInlining)]
+        /// <param name="name">The argument name.</param>
+        /// <param name="value">The current argument value.</param>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="ArgumentOutOfRangeException">Thrown with the specified parameters.</exception>
         [DoesNotReturn]
-        internal static void ThrowArgumentExceptionForsBitwiseEqualTo<T>(T value, T target, string name)
-            where T : unmanaged
+        public static void ThrowArgumentOutOfRangeException(string name, object value, string message)
         {
-            ThrowArgumentException(name, $"Parameter {name.ToAssertString()} ({typeof(T).ToTypeString()}) is not a bitwise match, was <{value.ToHexString()}> instead of <{target.ToHexString()}>");
+            throw new ArgumentOutOfRangeException(name, value, message);
         }
 
         /// <summary>
-        /// Throws an <see cref="ArgumentException"/> when <see cref="Guard.IsReferenceEqualTo{T}"/> fails.
+        /// Throws a new <see cref="ExternalException"/>.
         /// </summary>
-        /// <typeparam name="T">The type of input value being compared.</typeparam>
-        [MethodImpl(MethodImplOptions.NoInlining)]
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="ExternalException">Thrown with the specified parameter.</exception>
         [DoesNotReturn]
-        internal static void ThrowArgumentExceptionForIsReferenceEqualTo<T>(string name)
-            where T : class
+        public static void ThrowExternalException(string message)
         {
-            ThrowArgumentException(name, $"Parameter {name.ToAssertString()} ({typeof(T).ToTypeString()}) must be the same instance as the target object");
+            throw new ExternalException(message);
         }
 
         /// <summary>
-        /// Throws an <see cref="ArgumentException"/> when <see cref="Guard.IsReferenceNotEqualTo{T}"/> fails.
+        /// Throws a new <see cref="ExternalException"/>.
         /// </summary>
-        /// <typeparam name="T">The type of input value being compared.</typeparam>
-        [MethodImpl(MethodImplOptions.NoInlining)]
+        /// <param name="message">The argument name.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="ExternalException">Thrown with the specified parameters.</exception>
         [DoesNotReturn]
-        internal static void ThrowArgumentExceptionForIsReferenceNotEqualTo<T>(string name)
-            where T : class
+        public static void ThrowExternalException(string message, Exception innerException)
         {
-            ThrowArgumentException(name, $"Parameter {name.ToAssertString()} ({typeof(T).ToTypeString()}) must not be the same instance as the target object");
+            throw new ExternalException(message, innerException);
         }
 
         /// <summary>
-        /// Throws an <see cref="ArgumentException"/> when <see cref="Guard.IsTrue(bool,string)"/> fails.
+        /// Throws a new <see cref="ExternalException"/>.
         /// </summary>
-        [MethodImpl(MethodImplOptions.NoInlining)]
+        /// <param name="message">The argument name.</param>
+        /// <param name="error">The HRESULT of the errror to include.</param>
+        /// <exception cref="ExternalException">Thrown with the specified parameters.</exception>
         [DoesNotReturn]
-        internal static void ThrowArgumentExceptionForIsTrue(string name)
+        public static void ThrowExternalException(string message, int error)
         {
-            ThrowArgumentException(name, $"Parameter {name.ToAssertString()} must be true, was false");
+            throw new ExternalException(message, error);
         }
 
         /// <summary>
-        /// Throws an <see cref="ArgumentException"/> when <see cref="Guard.IsTrue(bool,string,string)"/> fails.
+        /// Throws a new <see cref="FormatException"/>.
         /// </summary>
-        [MethodImpl(MethodImplOptions.NoInlining)]
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="FormatException">Thrown with the specified parameter.</exception>
         [DoesNotReturn]
-        internal static void ThrowArgumentExceptionForIsTrue(string name, string message)
+        public static void ThrowFormatException(string message)
         {
-            ThrowArgumentException(name, $"Parameter {name.ToAssertString()} must be true, was false: {message.ToAssertString()}");
+            throw new FormatException(message);
         }
 
         /// <summary>
-        /// Throws an <see cref="ArgumentException"/> when <see cref="Guard.IsFalse(bool,string)"/> fails.
+        /// Throws a new <see cref="FormatException"/>.
         /// </summary>
-        [MethodImpl(MethodImplOptions.NoInlining)]
+        /// <param name="message">The message to include in the exception.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="FormatException">Thrown with the specified parameter.</exception>
         [DoesNotReturn]
-        internal static void ThrowArgumentExceptionForIsFalse(string name)
+        public static void ThrowFormatException(string message, Exception innerException)
         {
-            ThrowArgumentException(name, $"Parameter {name.ToAssertString()} must be false, was true");
+            throw new FormatException(message, innerException);
         }
 
         /// <summary>
-        /// Throws an <see cref="ArgumentException"/> when <see cref="Guard.IsFalse(bool,string,string)"/> fails.
+        /// Throws a new <see cref="InsufficientMemoryException"/>.
         /// </summary>
-        [MethodImpl(MethodImplOptions.NoInlining)]
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="InsufficientMemoryException">Thrown with the specified parameter.</exception>
         [DoesNotReturn]
-        internal static void ThrowArgumentExceptionForIsFalse(string name, string message)
+        public static void ThrowInsufficientMemoryException(string message)
         {
-            ThrowArgumentException(name, $"Parameter {name.ToAssertString()} must be false, was true: {message.ToAssertString()}");
+            throw new InsufficientMemoryException(message);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="InsufficientMemoryException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="InsufficientMemoryException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowInsufficientMemoryException(string message, Exception innerException)
+        {
+            throw new InsufficientMemoryException(message, innerException);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="InvalidDataException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="InvalidDataException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowInvalidDataException(string message)
+        {
+            throw new InvalidDataException(message);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="InvalidDataException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="InvalidDataException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowInvalidDataException(string message, Exception innerException)
+        {
+            throw new InvalidDataException(message, innerException);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="InvalidOperationException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="InvalidOperationException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowInvalidOperationException(string message)
+        {
+            throw new InvalidOperationException(message);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="InvalidOperationException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="InvalidOperationException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowInvalidOperationException(string message, Exception innerException)
+        {
+            throw new InvalidOperationException(message, innerException);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="LockRecursionException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="LockRecursionException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowLockRecursionException(string message)
+        {
+            throw new LockRecursionException(message);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="LockRecursionException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="LockRecursionException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowLockRecursionException(string message, Exception innerException)
+        {
+            throw new LockRecursionException(message, innerException);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="MissingFieldException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="MissingFieldException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowMissingFieldException(string message)
+        {
+            throw new MissingFieldException(message);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="MissingFieldException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="MissingFieldException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowMissingFieldException(string message, Exception innerException)
+        {
+            throw new MissingFieldException(message, innerException);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="MissingFieldException"/>.
+        /// </summary>
+        /// <param name="className">The target class being inspected.</param>
+        /// <param name="fieldName">The target field being retrieved.</param>
+        /// <exception cref="MissingFieldException">Thrown with the specified parameters.</exception>
+        [DoesNotReturn]
+        public static void ThrowMissingFieldException(string className, string fieldName)
+        {
+            throw new MissingFieldException(className, fieldName);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="MissingMemberException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="MissingMemberException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowMissingMemberException(string message)
+        {
+            throw new MissingMemberException(message);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="MissingMemberException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="MissingMemberException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowMissingMemberException(string message, Exception innerException)
+        {
+            throw new MissingMemberException(message, innerException);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="MissingMemberException"/>.
+        /// </summary>
+        /// <param name="className">The target class being inspected.</param>
+        /// <param name="memberName">The target member being retrieved.</param>
+        /// <exception cref="MissingMemberException">Thrown with the specified parameters.</exception>
+        [DoesNotReturn]
+        public static void ThrowMissingMemberException(string className, string memberName)
+        {
+            throw new MissingMemberException(className, memberName);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="MissingMethodException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="MissingMethodException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowMissingMethodException(string message)
+        {
+            throw new MissingMethodException(message);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="MissingMethodException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="MissingMethodException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowMissingMethodException(string message, Exception innerException)
+        {
+            throw new MissingMethodException(message, innerException);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="MissingMethodException"/>.
+        /// </summary>
+        /// <param name="className">The target class being inspected.</param>
+        /// <param name="methodName">The target method being retrieved.</param>
+        /// <exception cref="MissingMethodException">Thrown with the specified parameters.</exception>
+        [DoesNotReturn]
+        public static void ThrowMissingMethodException(string className, string methodName)
+        {
+            throw new MissingMethodException(className, methodName);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="NotSupportedException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="NotSupportedException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowNotSupportedException(string message)
+        {
+            throw new NotSupportedException(message);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="NotSupportedException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="NotSupportedException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowNotSupportedException(string message, Exception innerException)
+        {
+            throw new NotSupportedException(message, innerException);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="ObjectDisposedException"/>.
+        /// </summary>
+        /// <param name="objectName">The name of the disposed object.</param>
+        /// <exception cref="ObjectDisposedException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowObjectDisposedException(string objectName)
+        {
+            throw new ObjectDisposedException(objectName);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="ObjectDisposedException"/>.
+        /// </summary>
+        /// <param name="objectName">The name of the disposed object.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="ObjectDisposedException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowObjectDisposedException(string objectName, Exception innerException)
+        {
+            throw new ObjectDisposedException(objectName, innerException);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="ObjectDisposedException"/>.
+        /// </summary>
+        /// <param name="objectName">The name of the disposed object.</param>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="ObjectDisposedException">Thrown with the specified parameters.</exception>
+        [DoesNotReturn]
+        public static void ThrowObjectDisposedException(string objectName, string message)
+        {
+            throw new ObjectDisposedException(objectName, message);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="OperationCanceledException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="OperationCanceledException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowOperationCanceledException(string message)
+        {
+            throw new OperationCanceledException(message);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="OperationCanceledException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="OperationCanceledException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowOperationCanceledException(string message, Exception innerException)
+        {
+            throw new OperationCanceledException(message, innerException);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="OperationCanceledException"/>.
+        /// </summary>
+        /// <param name="token">The <see cref="CancellationToken"/> in use.</param>
+        /// <exception cref="OperationCanceledException">Thrown with the specified parameters.</exception>
+        [DoesNotReturn]
+        public static void ThrowOperationCanceledException(CancellationToken token)
+        {
+            throw new OperationCanceledException(token);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="OperationCanceledException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <param name="token">The <see cref="CancellationToken"/> in use.</param>
+        /// <exception cref="OperationCanceledException">Thrown with the specified parameters.</exception>
+        [DoesNotReturn]
+        public static void ThrowOperationCanceledException(string message, CancellationToken token)
+        {
+            throw new OperationCanceledException(message, token);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="OperationCanceledException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <param name="token">The <see cref="CancellationToken"/> in use.</param>
+        /// <exception cref="OperationCanceledException">Thrown with the specified parameters.</exception>
+        [DoesNotReturn]
+        public static void ThrowOperationCanceledException(string message, Exception innerException, CancellationToken token)
+        {
+            throw new OperationCanceledException(message, innerException, token);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="PlatformNotSupportedException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="PlatformNotSupportedException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowPlatformNotSupportedException(string message)
+        {
+            throw new PlatformNotSupportedException(message);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="PlatformNotSupportedException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="PlatformNotSupportedException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowPlatformNotSupportedException(string message, Exception innerException)
+        {
+            throw new PlatformNotSupportedException(message, innerException);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="SynchronizationLockException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="SynchronizationLockException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowSynchronizationLockException(string message)
+        {
+            throw new SynchronizationLockException(message);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="SynchronizationLockException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="SynchronizationLockException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowSynchronizationLockException(string message, Exception innerException)
+        {
+            throw new SynchronizationLockException(message, innerException);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="TimeoutException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="TimeoutException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowTimeoutException(string message)
+        {
+            throw new TimeoutException(message);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="TimeoutException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="TimeoutException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowTimeoutException(string message, Exception innerException)
+        {
+            throw new TimeoutException(message, innerException);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="UnauthorizedAccessException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="UnauthorizedAccessException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowUnauthorizedAccessException(string message)
+        {
+            throw new UnauthorizedAccessException(message);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="UnauthorizedAccessException"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="UnauthorizedAccessException">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowUnauthorizedAccessException(string message, Exception innerException)
+        {
+            throw new UnauthorizedAccessException(message, innerException);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="Win32Exception"/>.
+        /// </summary>
+        /// <param name="error">The Win32 error code associated with this exception.</param>
+        /// <exception cref="Win32Exception">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowWin32Exception(int error)
+        {
+            throw new Win32Exception(error);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="Win32Exception"/>.
+        /// </summary>
+        /// <param name="error">The Win32 error code associated with this exception.</param>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="Win32Exception">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowWin32Exception(int error, string message)
+        {
+            throw new Win32Exception(error, message);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="Win32Exception"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <exception cref="Win32Exception">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowWin32Exception(string message)
+        {
+            throw new Win32Exception(message);
+        }
+
+        /// <summary>
+        /// Throws a new <see cref="Win32Exception"/>.
+        /// </summary>
+        /// <param name="message">The message to include in the exception.</param>
+        /// <param name="innerException">The inner <see cref="Exception"/> to include.</param>
+        /// <exception cref="Win32Exception">Thrown with the specified parameter.</exception>
+        [DoesNotReturn]
+        public static void ThrowWin32Exception(string message, Exception innerException)
+        {
+            throw new Win32Exception(message, innerException);
         }
     }
 }
